+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # vNext
 
+* Fix: Fix setting in-app-includes from external properties (#1291)
+
 
 # 4.2.0