+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # vNext
 
+
 # 4.2.0
 
 * Fix: Remove experimental annotation for Attachment #1257