+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # vNext
 
 * Fix: Fix setting in-app-includes from external properties (#1291)
-
+* Fix: Fix JUL integration SDK name (#1293)
 
 # 4.2.0