+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix: Initialize Sentry in Logback appender when DSN is not set in XML config (#1296)
 * Fix: Fix JUL integration SDK name (#1293)
 * Feat: Activity tracing auto instrumentation
+* Ref: Separate user facing and protocol classes in the Performance feature (#1304)
 
 # 4.2.0