+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# vNext
+# 4.3.0
 
 * Fix: Fix setting in-app-includes from external properties (#1291)
 * Fix: Initialize Sentry in Logback appender when DSN is not set in XML config (#1296)